requirements.txt
crewai
crewai_tools
jupytext
litellm
notebook
python-dotenv
tavily-python
ipykernel